# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  response.Common:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.SuccessWithData-model_Comment:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.Comment'
      message:
        type: string
      status:
        type: string
    type: object
  response.SuccessWithData-model_Project:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.Project'
      message:
        type: string
      status:
        type: string
    type: object
  response.SuccessWithData-model_Section:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.Section'
      message:
        type: string
      status:
        type: string
    type: object
  response.SuccessWithData-model_Task:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/model.Task'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.SuccessWithPaginate-model_Project:
    properties:
      code:
        type: integer
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      status:
        type: string
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
  response.SuccessWithPaginate-model_Section:
    properties:
      code:
        type: integer
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/model.Section'
        type: array
      status:
        type: string
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  validation.AddGroupToProject:
    properties:
      group_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      project_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - group_id
    - project_id
    type: object
...
  validation.CreateComment:
    properties:
      body:
        example: fake comment
        type: string
      task_id:
        description: Добавил теги
        type: string
    required:
    - body
    - task_id
    type: object
  validation.CreateGroup:
    properties:
      project_id:
        type: string
      task_ids:
        description: Было []string
        items:
          type: string
        type: array
      title:
        example: fake name
        maxLength: 50
        type: string
    required:
    - project_id
    - title
    type: object
  validation.CreateProject:
    properties:
      title:
        example: fake name
        maxLength: 50
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/indrayyana/go-fiber-boilerplate/blob/main/LICENSE
  title: go-fiber-boilerplate API documentation
  version: 1.0.0
paths:
...
  /comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.CreateComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Comment task
      tags:
      - Comments
...
  /projects:
    get:
      consumes:
      - application/json
      description: Retrieve a list of projects of user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get projects by user
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create a new project with the provided details.
      parameters:
      - description: Project creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.CreateProject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /projects/{projectID}/sections:
    get:
      description: Retrieve all sections within a specific project.
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_Section'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sections of a project
      tags:
      - Sections
  /projects/add-group:
    post:
      consumes:
      - application/json
      description: Add a group to an existing project.
      parameters:
      - description: Add group to project request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.AddGroupToProject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a group to a project
      tags:
      - Projects
  /projects/section:
    post:
      consumes:
      - application/json
      description: Create a new section with the provided details.
      parameters:
      - description: Project section creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.CreateGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_Section'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new section
      tags:
      - Sections
  /sections:
    get:
      description: Retrieve all sections within a specific user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_UserSection'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sections of user
      tags:
      - Sections
  /sections/{sectionID}:
    delete:
      description: Delete a section by its unique ID.
      parameters:
      - description: Section ID
        in: path
        name: sectionID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete section by ID
      tags:
      - Sections
  /tasks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of tasks.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks of user
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the provided details.
      parameters:
      - description: Task creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Retrieve a list of users in a specific group.
      parameters:
      - description: Get users in group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.GetUsersInGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users in a group
      tags:
      - Tasks
  /tasks/{taskID}:
    delete:
      description: Delete a task by its unique ID.
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete task by ID
      tags:
      - Tasks
    get:
      description: Retrieve a task by its unique ID.
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - Tasks
  /tasks/{taskID}/reassign:
    put:
      consumes:
      - application/json
      description: Change the assignee of a task and notify via WebSocket.
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      - description: New assignee information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.ReassignTaskValidation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reassign task to a new user
      tags:
      - Tasks
  /tasks/{taskID}/users:
    get:
      description: Retrieve a list of users who have access to a specific task.
      parameters:
      - description: Task ID
        in: path
        name: taskID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-array_model_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users with access to a task
      tags:
      - Tasks
  /tasks/add-group:
    post:
      consumes:
      - application/json
      description: Добавление группы пользователей на задачу.
      parameters:
      - description: Add group to task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.AddGroupToTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a group to a task
      tags:
      - Tasks
  /user-groups:
    get:
      description: Retrieve a list of all user groups.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_UserGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all user groups
      tags:
      - UserGroups
    post:
      consumes:
      - application/json
      description: Create a new user group with the provided details.
      parameters:
      - description: User group creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.CreateUserGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithData-model_UserGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user group
      tags:
      - UserGroups
  /user-groups/add-user:
    post:
      consumes:
      - application/json
      description: Add a user to an existing group.
      parameters:
      - description: Add user to group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.AddUserToGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Common'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a user to a group
      tags:
      - UserGroups
  /user-groups/users:
    post:
      consumes:
      - application/json
      description: Retrieve a list of users in a specific group.
      parameters:
      - description: Get users in group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validation.GetUsersInGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessWithPaginate-model_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users in a group
      tags:
      - UserGroups
...
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"